requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
langchain-huggingface==0.0.1
langgraph==0.0.26
qdrant-client==1.7.0
streamlit==1.29.0
python-dotenv==1.0.0
youtube-transcript-api==0.6.1
duckduckgo-search==4.1.1
requests==2.31.0